refactor(saves): extract VersionsService class to saves/versions.py
VersionsService owns version history listing (list_file_versions) and
the rollback flow (rollback_to_version, _rollback_to_version_io,
_find_file_state). Rollback runs a pre-flight matrix sync, then
atomically replaces the currently-tracked server save with an older
version via download + PUT + tracked_save_id update.

Constructor takes a SaveService back-reference (TYPE_CHECKING guarded)
for the I/O orchestrators that still live on SaveService
(_do_download_save, _do_upload_save, _update_file_sync_state,
_record_own_upload, _get_rom_save_info, _sync_rom_saves). The
back-reference will be replaced with a SyncEngine reference once
SyncEngine is extracted.

Public methods on SaveService (list_file_versions, rollback_to_version)
remain as thin delegates so existing test and frontend callers work
unchanged. Private helpers are removed from SaveService where no
external caller depends on them.

Refs #257.
